Db_api/requirements.txt
-i https://pypi.org/simple
anyio==3.6.2 ; python_full_version >= '3.6.2'
click==8.1.3 ; python_version >= '3.7'
fastapi==0.95.1
h11==0.14.0 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
psycopg2-binary==2.9.6
pydantic==1.10.7 ; python_version >= '3.7'
sniffio==1.3.0 ; python_version >= '3.7'
starlette==0.26.1 ; python_version >= '3.7'
typing-extensions==4.5.0 ; python_version >= '3.7'
uvicorn==0.22.0
